refactor: 패키지 구조 재정의
