+++ src/main/java/com/programmers/java/calculator/CalculatorApplication.java
@@ -1,7 +1,7 @@
 package com.programmers.java.calculator;
 
-import com.programmers.java.calculator.calculator.PostfixExpressionCalculator;
 import com.programmers.java.calculator.controller.CalculatorController;
+import com.programmers.java.calculator.domain.calculator.PostfixExpressionCalculator;
 import com.programmers.java.calculator.repository.MemoryCalculatorRepository;
 import com.programmers.java.calculator.service.CalculatorService;
 import com.programmers.java.calculator.util.PostfixExpressionConverter;
