+++ src/test/java/com/programmers/java/calculator/CalculationTest.java
@@ -1,6 +1,6 @@
 package com.programmers.java.calculator;
 
-import com.programmers.java.calculator.calculator.PostfixExpressionCalculator;
+import com.programmers.java.calculator.domain.calculator.PostfixExpressionCalculator;
 import com.programmers.java.calculator.repository.MemoryCalculatorRepository;
 import com.programmers.java.calculator.service.CalculatorService;
 import com.programmers.java.calculator.util.PostfixExpressionConverter;
